move ImplStrategy to a value_trait module
