+++ derive_macros/src/stream/stream_newtype.rs
@@ -1,12 +1,12 @@
 use syn::{spanned::Spanned, Field, Ident, Path};
 
-use crate::derive::field_codegen;
+use crate::value_trait::field_codegen;
 use crate::{
     attr,
-    derive::ImplStrategy,
     index::{quote_optional_index, Index},
     label::{quote_optional_label, Label},
     tag::quote_optional_tag,
+    value_trait::ImplStrategy,
 };
 
 pub(crate) fn stream_newtype<B>(
